changes for no two component in one file
